LayoutTests/crypto/subtle/aes-cbc-encrypt-decrypt-with-padding-expected.txt

Test AES-CBC encrypt and decrypt functions on a plaintext that is not a multiple of block size in length.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a raw AES key from string literal...
PASS key.type is 'secret'
PASS key.extractable is true
PASS key.algorithm.name is 'aes-cbc'
PASS key.algorithm.length is 128
PASS key.usages is ['encrypt', 'decrypt']
Using the key to encrypt plaintext...
    = [63 01 99 c5 f2 02 cc 71 67 bb 84 c6 c7 2b 34 9d]
Decrypting it back...
PASS new Uint8Array(decryptionResult) is plaintext
PASS successfullyParsed is true

TEST COMPLETE
